requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asyncpg==0.29.0; python_full_version >= '3.8.0'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
email-validator==2.2.0; python_version >= '3.8'
fastapi==0.111.1; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
mailtrap==2.0.1; python_version >= '3.6'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pyyaml==6.0.1
requests==2.32.3; python_version >= '3.0'
rich==13.7.1; python_full_version >= '3.7.0'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
stripe==10.5.0; python_version >= '3.6'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.7'
urllib3==2.2.2; python_version >= '3.8'
uvicorn[standard]==0.30.3; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
